scalable_export_worker/requirements.txt
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
gevent==20.9.0
google-api-core==1.23.0
google-auth==1.24.0
google-cloud-bigquery==2.6.1
google-cloud-core==1.4.4
google-cloud-logging==2.0.1
google-cloud-pubsub==2.2.0
google-cloud-storage==1.34.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpc-google-iam-v1==0.12.3
grpcio==1.34.0
idna==2.10
langdetect==1.0.8
libcst==0.3.15
mypy-extensions==0.4.3
numpy==1.19.4
pandas==1.1.5
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
zope.event==4.5.0
zope.interface==5.2.0
